Add a model module
